Pull request to update forked repo

+++ src/rcpp_branch_matrix.cpp
@@ -18,19 +18,20 @@ arma::sp_mat rcpp_branch_matrix(Rcpp::List x) {
   std::size_t curr_node;
   std::size_t curr_branch;
   // main processing
-  for (std::size_t curr_tip=0; curr_tip<tip_labels.size(); ++curr_tip) {
+  for (std::size_t curr_tip = 0;
+       curr_tip < static_cast<std::size_t>(tip_labels.size()); ++curr_tip) {
     // set variables to be the tip corresponding to the i'th species
     curr_node = (curr_tip+1);
     // loop through branches until the curr_pos is pointing to the base
     while (true) {
-      // find the element where curr_node is at the top of the branch 
-      curr_pos = std::find(edge_column_2.begin(), edge_column_2.end(), 
+      // find the element where curr_node is at the top of the branch
+      curr_pos = std::find(edge_column_2.begin(), edge_column_2.end(),
                            curr_node);
       // break if the base has been found
       if (curr_pos == edge_column_2.end()) {
         break;
       }
-      // calculate the branch number for the node (ie. the index of the 
+      // calculate the branch number for the node (ie. the index of the
       // of the node
       curr_branch = std::distance(edge_column_2.begin(), curr_pos);
       // store the branch number and the tip number
